Fix some memory leaks
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1864 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/ConstraintManager/OclViolation.h
@@ -40,8 +40,8 @@ public:
 			}
 		}
 		clear();
-		int lenx = vect.size();
-		for (int kk=0; kk<lenx; kk++)
+		size_t lenx = vect.size();
+		for (size_t kk=0; kk<lenx; kk++)
 		{
 			IUnknown *ptr = vect[kk];
 			if (ptr)
